[#87854] Some code cleanup after merge with main

+++ backend/src/test/java/nu/ndw/nls/accessibilitymap/backend/municipality/mapper/MunicipalityFeatureMapperTest.java
@@ -75,7 +75,7 @@ class MunicipalityFeatureMapperTest {
 
         var propertiesJson = new MunicipalityPropertiesJson(
                 municipality.name(),
-                (int) municipality.searchDistanceInMetres(),
+                municipality.searchDistanceInMetres(),
                 List.of(boundsStart, boundsEnd),
                 dateLastCheck);
